# ubkg-api-spec.yaml
openapi: 3.0.3
info:
  title: UBKG API
  description: The UBKG API provides REST endpoints that return results from queries against an instance of the [Unified Biomedical Knowledge Graph (UBKG)](https://ubkg.docs.xconsortia.org/). To prevent exceeding service timeouts or payload limits, some endpoints are throttled in both execution time and in the size of its response payload--e.g., to a maximum execution time of 15 seconds and a maximum payload of 9 MB.
  version: "2.0"
  contact:
    name: GitHub repository
    url: https://github.com/x-atlas-consortia/ubkg-api
servers:
  - url: https://ontology.api.hubmapconsortium.org/
    description: Production server
  - url: https://ontology-api.dev.hubmapconsortium.org/
    description: Development server
paths:
...
  /node_types/counts:
    get:
      summary: Returns counts of nodes in the database by node type (label).
      operationId: node_type_count_get
      responses:
        '200':
          description: List of counts by node type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeTypeCount'
        '404':
          description: No nodes with label
        '5XX':
          description: Unknown error
  /node_types/{node_type}/counts:
    get:
      summary: Returns counts of nodes in the database for a specified node type (label).
      operationId: node_type_count_node_id_get
      parameters:
        - name: node_type
          in: path
          required: true
          description: a node type
          schema:
            type: string
            example: Code
      responses:
        '200':
          description: List of counts by node type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeTypeCount'
        '404':
          description: No nodes with label
        '5XX':
          description: Unknown error
  /node_types/counts_by_sab:
    get:
      summary: This endpoint exists as a convenience to explain a limitation; a custom 400 message will explain the issue and suggest alternatives.
      operationId: node_type_count_by_sab_get
      parameters:
        - name: sab
          in: query
          required: false
          description: The set of sources (SABs) to which to limit counts by node_type.
          schema:
            type: string
            example: NCI
      responses:
        '400':
          description: The response to this endpoint is likely to exceed timeout and so will not be attempted. Execute the node_types/counts_by_sab/(node_type) endpoint with the name of a node type (e.g., Codes). To obtain names of node types, execute the node_types/counts endpoint.
        '5XX':
          description: Unknown error
  /node_types/{node_type}/counts_by_sab:
    get:
      summary: Returns counts of nodes in the database for a specified node type (label), grouped by source (SAB), subject to constraints specified in parameters.
      operationId: node_type_count_by_sab_node_type_get
      parameters:
        - name: node_type
          in: path
          required: true
          description: a node type
          schema:
            type: string
            example: Code
        - name: sab
          in: query
          required: false
          description: The set of sources (SABs) to which to limit counts by node_type. Can be either a list of values delimited with URL-encoded commas (e.g., ?sab=SAB1%2CSAB2%2SAB3) or with individual key-value pairs (e.g., ?sab=SAB1&sab=SAB2&SAB=SAB3).
          schema:
            type: string
            example: NCI
      responses:
        '200':
          description: List of counts by node type, grouped by source, subject to constraints specified in parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeTypeSabCount'
        '400':
          description: invalid parameter
        '404':
          description: No nodes with label
        '5XX':
          description: Unknown error
...
  /sabs/codes/counts:
    get:
      operationId: sabs_codes_counts_get
      summary: Returns a set of sources (SABs), including counts of the codes associated with the sources, subject to constraints from parameters
      parameters:
        - name: skip
          in: query
          required: false
          description: the number of Codes to skip in the returned set. The value must be non-negative. The default value is 0.
          schema:
            type: string
            example: 0
        - name: limit
          in: query
          required: false
          description: the maximum number of Codes to return. The value must be non-negative. The default value is the maximum number of rows specified by configuration.
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: Returns information on a set of sources, subject to constraints from parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SabCodeCounts'
        '400':
          description: invalid parameter name; non-numeric parameter value; negative parameter value
        '404':
          description: No Codes for the specified parameters
        '5XX':
          description: Unknown error
  /sabs/{sab}/codes/counts:
    get:
      operationId: sabs_codes_counts_sab_get
      summary: Returns information for a specified source (SAB), including counts of the codes associated with the source, subject to constraints from parameters
      parameters:
        - name: sab
          in: path
          required: true
          description: A source (SAB) to which to limit the response.
          schema:
            type: string
            example: UBERON
        - name: skip
          in: query
          required: false
          description: the number of Codes to skip in the returned set. The value must be non-negative. The default value is 0.
          schema:
            type: string
            example: 0
        - name: limit
          in: query
          required: false
          description: the maximum number of Codes to return. The value must be non-negative. The default value is the maximum number of rows specified by configuration.
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: Returns information on a source, subject to constraints from parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SabCodeDetails'
        '400':
          description: invalid parameter name; non-numeric parameter value; negative parameter value
        '404':
          description: No Codes for the specified parameters
        '5XX':
          description: Unknown error
  /sabs/codes/details:
    get:
      operationId: sabs_codes_details_get
      summary: This endpoint exists as a convenience to explain a limitation; a custom 400 message will explain the issue and suggest alternatives.
      responses:
        '400':
          description: The response to this endpoint cannot be run for all SABs because of memory limitations. Execute the /sabs/(sab)/codes/details endpoint with the identifier for a SAB. Execute the /sabs/codes/counts endpoint for a list of all SABs in the UBKG.
        '5XX':
          description: Unknown error
  /sabs/{sab}/codes/details:
    get:
      operationId: sabs_codes_details_sab_get
      summary: Returns including details for codes associated with a specified source (SAB), subject to constraints from parameters
      parameters:
        - name: sab
          in: path
          required: true
          description: A source (SAB) to which to limit the response.
          schema:
            type: string
            example: UBERON
        - name: skip
          in: query
          required: false
          description: the number of Codes to skip in the returned set. The value must be non-negative. The default value is 0.
          schema:
            type: string
            example: 0
        - name: limit
          in: query
          required: false
          description: the maximum number of Codes to return. The value must be non-negative. The default value is the maximum number of rows specified by configuration.
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: Returns information on a set of sources and their associated Codes, subject to constraints from parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SabCodeDetails'
        '400':
          description: invalid parameter name; non-numeric parameter value; negative parameter value
        '404':
          description: No Codes for the specified parameters
        '5XX':
          description: Unknown error
  /sabs/term_types:
    get:
      operationId: sabs_term_types_get
      summary: This endpoint exists as a convenience to explain a limitation; a custom 400 message will explain the issue and suggest alternatives.
      responses:
        '400':
          description: The response to this endpoint cannot be run for all SABs because of memory limitations. Execute the /sabs/(sab)/codes/term_types endpoint with the identifier for a SAB. Execute the /sabs/codes/counts endpoint for a list of all SABs in the UBKG.
        '5XX':
          description: Unknown error
  /sabs/{sab}/term_types:
    get:
      operationId: sabs_sab_term_types_get
      summary: Returns the list of term types (types of relationship) for relationships between the nodes that are defined by the specified source (SAB)
      parameters:
        - name: sab
          in: path
          required: true
          description: A source (SAB) to which to limit the response.
          schema:
            type: string
            example: UBERON
        - name: skip
          in: query
          required: false
          description: the number of Codes to skip in the returned set. The value must be non-negative. The default value is 0.
          schema:
            type: string
            example: 0
        - name: limit
          in: query
          required: false
          description: the maximum number of Codes to return. The value must be non-negative. The default value is the maximum number of rows specified by configuration.
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: Returns information on a set of sources and their associated Codes, subject to constraints from parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SabTermTypes'
        '400':
          description: invalid parameter name; non-numeric parameter value; negative parameter value
        '404':
          description: No term types for the specified parameters
        '5XX':
          description: Unknown error
...
components:
  schemas:
...
    NodeTypeCount: # Schema name
      type: object
      description: counts by node type
      properties:
        node_types:
          type: array
          items:
            type: object
            properties:
              total_count:
                type: integer
                description: Count of nodes with any of the labels in node_types. Equal to the total count of nodes in the database if no node_type was specified.
                example: 1650946
              node_type:
                type: object
                properties:
                  label:
                    type: string
                    description: the label for the node type
                    example: Code
                  count:
                    type: integer
                    description: count of nodes with the label
                    example: 16590946
    NodeTypeSabCount: # Schema name
      type: object
      description: counts by node type, grouped by SAB
      properties:
        node_types:
          type: array
          items:
            type: object
            properties:
              total_count:
                type: integer
                description: Count of nodes with any of the labels in node_types. Equal to the total count of nodes in the database if no filters were applied.
                example: 170947
              node_type:
                type: object
                properties:
                  label:
                    type: string
                    description: the label for the node type
                    example: Code
                  count:
                    type: integer
                    description: count of nodes with the label
                    example: 170947
                  sabs:
                    type: array
                    description: array of sources (SABs) for nodes with the specified node types
                    items:
                      type: object
                      properties:
                        sab:
                          type: string
                          description: source name
                          example: NCI
                        count:
                          type: integer
                          description: count of nodes from the SAB with the specified node type
                          example: 170947
...
    SabCodeCounts: #Schema name
      type: object
      description: A set of information on a source (SAB), including counts by codes.
      properties:
        sabs:
          description: list of SABs
          type: array
          items:
            type: object
            description: a source (SAB)
            properties:
              sab:
                description: source identifier
                type: string
                example: UBERON
              count:
                description: count of the codes in the SAB
                type: string
                example: 14314
              position:
                description: ordinal position of the sab in the endpoint's return, with respect to the _skip_ parameter
                type: integer
                example: 1
    SabCodeDetails: #Schema name
      type: object
      description: A set of information on the Codes of a source (SAB)
      properties:
        codes:
          description: list of Codes
          type: array
          items:
            type: object
            description: a Code in a SAB
            properties:
              codeid:
                description: the CodeID of the Code, in format _SAB_:_CODE_.
                type: string
                example: "UBERON:0000000"
              position:
                description: the position of the Code in the endpoint's return, with respect to the _skip_ parameter
                type: integer
                example: 1
              terms:
                description: set of terms associated with the Code.
                type: array
                items:
                  type: object
                  description: a term for a Code
                  properties:
                    term:
                      description: the term string
                      type: string
                      example: processual entity
                    term_type:
                      description: the term type for the term
                      type: string
                      example: PT
    SabTermTypes: #Schema name
      type: object
      description: A set of term types (relationship types) for relationships that originate from the Code nodes that are defined by a source (SAB).
      properties:
        sab:
          type: string
          description: identifier for a source
          example: SNOMEDCT_US
        term_types:
          type: array
          description: array of term types
          items:
            type: string
            description: the term type
            example: PT